feat(#37): Renamed builders into factory
