+++ docs/specs/#37.md
@@ -109,11 +109,12 @@ block
 - [x] Mappers for the entities and BL models
   - [x] Person and PersonEntity
   - [x] Recipe and RecipeEntity
-- [ ] Entering names must be have max length of 100
-- [ ] Entering recipe names must be have max length of 255
-- [ ] Change activity level into table and remove the hardcoded mapping.
-- [ ] The script to generate test data must be updated to generate data into the database instead of YAML files.
+- [x] Entering names must be have max length of 100
+- [x] Entering recipe names must be have max length of 255
+- [x] The script to generate test data must be updated to generate data into the database instead of YAML files.
+- [x] Rename builders into factories
 - [ ] "Do you want to save the recipe 'Skyr' to disk? (Y/n)" must be renamed.
+- [ ] Change activity level into table and remove the hardcoded mapping.
 - [ ] Check if the exception handling is well done
 - [ ] Check if further tests must be written
 - [ ] Are there license conflicts for new dependencies?
