+++ src/nutrition/command_line_handler.py
@@ -4,7 +4,7 @@ import questionary
 from rich.table import Table
 from common.terminal import terminal
 from nutrition.repository import Repository
-from nutrition.openfoodfacts_api_builder import api_client
+from nutrition.openfoodfacts_api_factory import api_client
 
 log = logging.getLogger(__name__)
